backend/requirements.txt
flask
flask-cors
torch
torchvision
opencv-python-headless
numpy
matplotlib
PyJWT
bcrypt
Pillow
gunicorn
